test: split skill detail page suite
